Fix column resizing bug
[PROBLEM]
When trying to resize the last column, the width was originally set to
X. If we tried to make the column smaller and it would cause the columns
to no longer take the available container width then the width of the
column would be set back to X. This meant no change was occurring in the
column observable so the column would remain at the width Y.

[SOLUTION]
Always update column observables no matter what. It will trigger updates
even if nothing changes.